+++ src/meta/src/controller/catalog/mod.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod alter_op;
+mod create_op;
 mod get_op;
 mod list_op;
 mod test;
+mod util;
 
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::iter;
@@ -73,6 +76,7 @@ use super::utils::{
     check_subscription_name_duplicate, get_fragment_ids_by_jobs, get_internal_tables_by_id,
     rename_relation, rename_relation_refer,
 };
+use crate::controller::catalog::util::update_internal_tables;
 use crate::controller::utils::{
     build_relation_group_for_delete, check_connection_name_duplicate,
     check_database_name_duplicate, check_function_signature_duplicate,
@@ -148,77 +152,6 @@ impl CatalogController {
         Ok(catalog_controller)
     }
 
-    async fn init(&self) -> MetaResult<()> {
-        self.table_catalog_cdc_table_id_update().await?;
-        Ok(())
-    }
-
-    /// Fill in the `cdc_table_id` field for Table with empty `cdc_table_id` and parent Source job.
-    /// NOTES: We assume Table with a parent Source job is a CDC table
-    async fn table_catalog_cdc_table_id_update(&self) -> MetaResult<()> {
-        let inner = self.inner.read().await;
-        let txn = inner.db.begin().await?;
-
-        // select Tables which cdc_table_id is empty and has a parent Source job
-        let table_and_source_id: Vec<(TableId, String, SourceId)> = Table::find()
-            .join(JoinType::InnerJoin, table::Relation::ObjectDependency.def())
-            .join(
-                JoinType::InnerJoin,
-                object_dependency::Relation::Source.def(),
-            )
-            .select_only()
-            .columns([table::Column::TableId, table::Column::Definition])
-            .columns([source::Column::SourceId])
-            .filter(
-                table::Column::TableType.eq(TableType::Table).and(
-                    table::Column::CdcTableId
-                        .is_null()
-                        .or(table::Column::CdcTableId.eq("")),
-                ),
-            )
-            .into_tuple()
-            .all(&txn)
-            .await?;
-
-        // return directly if the result set is empty.
-        if table_and_source_id.is_empty() {
-            return Ok(());
-        }
-
-        info!(table_and_source_id = ?table_and_source_id, "cdc table with empty cdc_table_id");
-
-        let mut cdc_table_ids = HashMap::new();
-        for (table_id, definition, source_id) in table_and_source_id {
-            match extract_external_table_name_from_definition(&definition) {
-                None => {
-                    tracing::warn!(
-                        table_id = table_id,
-                        definition = definition,
-                        "failed to extract cdc table name from table definition.",
-                    )
-                }
-                Some(external_table_name) => {
-                    cdc_table_ids.insert(
-                        table_id,
-                        build_cdc_table_id(source_id as u32, &external_table_name),
-                    );
-                }
-            }
-        }
-
-        for (table_id, cdc_table_id) in cdc_table_ids {
-            table::ActiveModel {
-                table_id: Set(table_id as _),
-                cdc_table_id: Set(Some(cdc_table_id)),
-                ..Default::default()
-            }
-            .update(&txn)
-            .await?;
-        }
-        txn.commit().await?;
-        Ok(())
-    }
-
     /// Used in `NotificationService::subscribe`.
     /// Need to pay attention to the order of acquiring locks to prevent deadlock problems.
     pub async fn get_inner_read_guard(&self) -> RwLockReadGuard<'_, CatalogControllerInner> {
@@ -269,68 +202,6 @@ impl CatalogController {
 }
 
 impl CatalogController {
-    pub(crate) async fn create_object(
-        txn: &DatabaseTransaction,
-        obj_type: ObjectType,
-        owner_id: UserId,
-        database_id: Option<DatabaseId>,
-        schema_id: Option<SchemaId>,
-    ) -> MetaResult<object::Model> {
-        let active_db = object::ActiveModel {
-            oid: Default::default(),
-            obj_type: Set(obj_type),
-            owner_id: Set(owner_id),
-            schema_id: Set(schema_id),
-            database_id: Set(database_id),
-            initialized_at: Default::default(),
-            created_at: Default::default(),
-            initialized_at_cluster_version: Set(Some(current_cluster_version())),
-            created_at_cluster_version: Set(Some(current_cluster_version())),
-        };
-        Ok(active_db.insert(txn).await?)
-    }
-
-    pub async fn create_database(&self, db: PbDatabase) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = db.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        check_database_name_duplicate(&db.name, &txn).await?;
-
-        let db_obj = Self::create_object(&txn, ObjectType::Database, owner_id, None, None).await?;
-        let mut db: database::ActiveModel = db.into();
-        db.database_id = Set(db_obj.oid);
-        let db = db.insert(&txn).await?;
-
-        let mut schemas = vec![];
-        for schema_name in iter::once(DEFAULT_SCHEMA_NAME).chain(SYSTEM_SCHEMAS) {
-            let schema_obj =
-                Self::create_object(&txn, ObjectType::Schema, owner_id, Some(db_obj.oid), None)
-                    .await?;
-            let schema = schema::ActiveModel {
-                schema_id: Set(schema_obj.oid),
-                name: Set(schema_name.into()),
-            };
-            let schema = schema.insert(&txn).await?;
-            schemas.push(ObjectModel(schema, schema_obj).into());
-        }
-        txn.commit().await?;
-
-        let mut version = self
-            .notify_frontend(
-                NotificationOperation::Add,
-                NotificationInfo::Database(ObjectModel(db, db_obj).into()),
-            )
-            .await;
-        for schema in schemas {
-            version = self
-                .notify_frontend(NotificationOperation::Add, NotificationInfo::Schema(schema))
-                .await;
-        }
-
-        Ok(version)
-    }
-
     pub async fn drop_database(
         &self,
         database_id: DatabaseId,
@@ -445,48 +316,6 @@ impl CatalogController {
         ))
     }
 
-    pub async fn get_object_database_id(&self, object_id: ObjectId) -> MetaResult<DatabaseId> {
-        let inner = self.inner.read().await;
-        let (database_id,): (Option<DatabaseId>,) = Object::find_by_id(object_id)
-            .select_only()
-            .select_column(object::Column::DatabaseId)
-            .into_tuple()
-            .one(&inner.db)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("object", object_id))?;
-        Ok(database_id.ok_or_else(|| anyhow!("object has no database id: {object_id}"))?)
-    }
-
-    pub async fn create_schema(&self, schema: PbSchema) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = schema.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, schema.database_id as _, &txn).await?;
-        check_schema_name_duplicate(&schema.name, schema.database_id as _, &txn).await?;
-
-        let schema_obj = Self::create_object(
-            &txn,
-            ObjectType::Schema,
-            owner_id,
-            Some(schema.database_id as _),
-            None,
-        )
-        .await?;
-        let mut schema: schema::ActiveModel = schema.into();
-        schema.schema_id = Set(schema_obj.oid);
-        let schema = schema.insert(&txn).await?;
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Add,
-                NotificationInfo::Schema(ObjectModel(schema, schema_obj).into()),
-            )
-            .await;
-        Ok(version)
-    }
-
     pub async fn drop_schema(
         &self,
         schema_id: SchemaId,
@@ -547,42 +376,6 @@ impl CatalogController {
             )
             .await;
         Ok(version)
-    }
-
-    pub async fn create_subscription_catalog(
-        &self,
-        pb_subscription: &mut PbSubscription,
-    ) -> MetaResult<()> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-
-        ensure_user_id(pb_subscription.owner as _, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_subscription.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_subscription.schema_id as _, &txn).await?;
-        check_subscription_name_duplicate(pb_subscription, &txn).await?;
-
-        let obj = Self::create_object(
-            &txn,
-            ObjectType::Subscription,
-            pb_subscription.owner as _,
-            Some(pb_subscription.database_id as _),
-            Some(pb_subscription.schema_id as _),
-        )
-        .await?;
-        pb_subscription.id = obj.oid as _;
-        let subscription: subscription::ActiveModel = pb_subscription.clone().into();
-        Subscription::insert(subscription).exec(&txn).await?;
-
-        // record object dependency.
-        ObjectDependency::insert(object_dependency::ActiveModel {
-            oid: Set(pb_subscription.dependent_table_id as _),
-            used_by: Set(pb_subscription.id as _),
-            ..Default::default()
-        })
-        .exec(&txn)
-        .await?;
-        txn.commit().await?;
-        Ok(())
     }
 
     pub async fn finish_create_subscription_catalog(&self, subscription_id: u32) -> MetaResult<()> {
@@ -677,119 +470,6 @@ impl CatalogController {
             .await?;
         txn.commit().await?;
         Ok(())
-    }
-
-    async fn list_object_dependencies(
-        &self,
-        include_creating: bool,
-    ) -> MetaResult<Vec<PbObjectDependencies>> {
-        let inner = self.inner.read().await;
-
-        let dependencies: Vec<(ObjectId, ObjectId)> = {
-            let filter = if include_creating {
-                Expr::value(true)
-            } else {
-                streaming_job::Column::JobStatus.eq(JobStatus::Created)
-            };
-            ObjectDependency::find()
-                .select_only()
-                .columns([
-                    object_dependency::Column::Oid,
-                    object_dependency::Column::UsedBy,
-                ])
-                .join(
-                    JoinType::InnerJoin,
-                    object_dependency::Relation::Object1.def(),
-                )
-                .join(JoinType::InnerJoin, object::Relation::StreamingJob.def())
-                .filter(filter)
-                .into_tuple()
-                .all(&inner.db)
-                .await?
-        };
-        let mut obj_dependencies = dependencies
-            .into_iter()
-            .map(|(oid, used_by)| PbObjectDependencies {
-                object_id: used_by as _,
-                referenced_object_id: oid as _,
-            })
-            .collect_vec();
-
-        let view_dependencies: Vec<(ObjectId, ObjectId)> = ObjectDependency::find()
-            .select_only()
-            .columns([
-                object_dependency::Column::Oid,
-                object_dependency::Column::UsedBy,
-            ])
-            .join(
-                JoinType::InnerJoin,
-                object_dependency::Relation::Object1.def(),
-            )
-            .join(JoinType::InnerJoin, object::Relation::View.def())
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-
-        obj_dependencies.extend(view_dependencies.into_iter().map(|(view_id, table_id)| {
-            PbObjectDependencies {
-                object_id: table_id as _,
-                referenced_object_id: view_id as _,
-            }
-        }));
-
-        let sink_dependencies: Vec<(SinkId, TableId)> = {
-            let filter = if include_creating {
-                sink::Column::TargetTable.is_not_null()
-            } else {
-                streaming_job::Column::JobStatus
-                    .eq(JobStatus::Created)
-                    .and(sink::Column::TargetTable.is_not_null())
-            };
-            Sink::find()
-                .select_only()
-                .columns([sink::Column::SinkId, sink::Column::TargetTable])
-                .join(JoinType::InnerJoin, sink::Relation::Object.def())
-                .join(JoinType::InnerJoin, object::Relation::StreamingJob.def())
-                .filter(filter)
-                .into_tuple()
-                .all(&inner.db)
-                .await?
-        };
-        obj_dependencies.extend(sink_dependencies.into_iter().map(|(sink_id, table_id)| {
-            PbObjectDependencies {
-                object_id: table_id as _,
-                referenced_object_id: sink_id as _,
-            }
-        }));
-
-        let subscription_dependencies: Vec<(SubscriptionId, TableId)> = {
-            let filter = if include_creating {
-                subscription::Column::DependentTableId.is_not_null()
-            } else {
-                subscription::Column::SubscriptionState
-                    .eq(Into::<i32>::into(SubscriptionState::Created))
-                    .and(subscription::Column::DependentTableId.is_not_null())
-            };
-            Subscription::find()
-                .select_only()
-                .columns([
-                    subscription::Column::SubscriptionId,
-                    subscription::Column::DependentTableId,
-                ])
-                .join(JoinType::InnerJoin, subscription::Relation::Object.def())
-                .filter(filter)
-                .into_tuple()
-                .all(&inner.db)
-                .await?
-        };
-        obj_dependencies.extend(subscription_dependencies.into_iter().map(
-            |(subscription_id, table_id)| PbObjectDependencies {
-                object_id: subscription_id as _,
-                referenced_object_id: table_id as _,
-            },
-        ));
-
-        Ok(obj_dependencies)
     }
 
     pub async fn has_any_streaming_jobs(&self) -> MetaResult<bool> {
@@ -1205,100 +885,6 @@ impl CatalogController {
         Ok(true)
     }
 
-    pub async fn create_source(
-        &self,
-        mut pb_source: PbSource,
-    ) -> MetaResult<(SourceId, NotificationVersion)> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_source.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_source.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_source.schema_id as _, &txn).await?;
-        check_relation_name_duplicate(
-            &pb_source.name,
-            pb_source.database_id as _,
-            pb_source.schema_id as _,
-            &txn,
-        )
-        .await?;
-
-        // handle secret ref
-        let secret_ids = get_referred_secret_ids_from_source(&pb_source)?;
-        let connection_ids = get_referred_connection_ids_from_source(&pb_source);
-
-        let source_obj = Self::create_object(
-            &txn,
-            ObjectType::Source,
-            owner_id,
-            Some(pb_source.database_id as _),
-            Some(pb_source.schema_id as _),
-        )
-        .await?;
-        let source_id = source_obj.oid;
-        pb_source.id = source_id as _;
-        let source: source::ActiveModel = pb_source.clone().into();
-        Source::insert(source).exec(&txn).await?;
-
-        // add secret dependency
-        let dep_relation_ids = secret_ids.iter().chain(connection_ids.iter());
-        if !secret_ids.is_empty() || !connection_ids.is_empty() {
-            ObjectDependency::insert_many(dep_relation_ids.map(|id| {
-                object_dependency::ActiveModel {
-                    oid: Set(*id as _),
-                    used_by: Set(source_id as _),
-                    ..Default::default()
-                }
-            }))
-            .exec(&txn)
-            .await?;
-        }
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend_relation_info(
-                NotificationOperation::Add,
-                PbRelationInfo::Source(pb_source),
-            )
-            .await;
-        Ok((source_id, version))
-    }
-
-    pub async fn create_function(
-        &self,
-        mut pb_function: PbFunction,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_function.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_function.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_function.schema_id as _, &txn).await?;
-        check_function_signature_duplicate(&pb_function, &txn).await?;
-
-        let function_obj = Self::create_object(
-            &txn,
-            ObjectType::Function,
-            owner_id,
-            Some(pb_function.database_id as _),
-            Some(pb_function.schema_id as _),
-        )
-        .await?;
-        pb_function.id = function_obj.oid as _;
-        let function: function::ActiveModel = pb_function.clone().into();
-        Function::insert(function).exec(&txn).await?;
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Add,
-                NotificationInfo::Function(pb_function),
-            )
-            .await;
-        Ok(version)
-    }
-
     pub async fn drop_function(&self, function_id: FunctionId) -> MetaResult<NotificationVersion> {
         let inner = self.inner.write().await;
         let txn = inner.db.begin().await?;
@@ -1338,89 +924,6 @@ impl CatalogController {
                 }),
             )
             .await;
-        Ok(version)
-    }
-
-    pub async fn create_secret(
-        &self,
-        mut pb_secret: PbSecret,
-        secret_plain_payload: Vec<u8>,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_secret.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_secret.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_secret.schema_id as _, &txn).await?;
-        check_secret_name_duplicate(&pb_secret, &txn).await?;
-
-        let secret_obj = Self::create_object(
-            &txn,
-            ObjectType::Secret,
-            owner_id,
-            Some(pb_secret.database_id as _),
-            Some(pb_secret.schema_id as _),
-        )
-        .await?;
-        pb_secret.id = secret_obj.oid as _;
-        let secret: secret::ActiveModel = pb_secret.clone().into();
-        Secret::insert(secret).exec(&txn).await?;
-
-        txn.commit().await?;
-
-        // Notify the compute and frontend node plain secret
-        let mut secret_plain = pb_secret;
-        secret_plain.value.clone_from(&secret_plain_payload);
-
-        LocalSecretManager::global().add_secret(secret_plain.id, secret_plain_payload);
-        self.env
-            .notification_manager()
-            .notify_compute_without_version(Operation::Add, Info::Secret(secret_plain.clone()));
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Add,
-                NotificationInfo::Secret(secret_plain),
-            )
-            .await;
-
-        Ok(version)
-    }
-
-    pub async fn alter_secret(
-        &self,
-        pb_secret: PbSecret,
-        secret_plain_payload: Vec<u8>,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_secret.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_secret.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_secret.schema_id as _, &txn).await?;
-
-        ensure_object_id(ObjectType::Secret, pb_secret.id as _, &txn).await?;
-        let secret: secret::ActiveModel = pb_secret.clone().into();
-        Secret::update(secret).exec(&txn).await?;
-
-        txn.commit().await?;
-
-        // Notify the compute and frontend node plain secret
-        let mut secret_plain = pb_secret;
-        secret_plain.value.clone_from(&secret_plain_payload);
-
-        LocalSecretManager::global().update_secret(secret_plain.id, secret_plain_payload);
-        self.env
-            .notification_manager()
-            .notify_compute_without_version(Operation::Update, Info::Secret(secret_plain.clone()));
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::Secret(secret_plain),
-            )
-            .await;
-
         Ok(version)
     }
 
@@ -1469,80 +972,6 @@ impl CatalogController {
         Ok(version)
     }
 
-    pub async fn create_connection(
-        &self,
-        mut pb_connection: PbConnection,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_connection.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_connection.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_connection.schema_id as _, &txn).await?;
-        check_connection_name_duplicate(&pb_connection, &txn).await?;
-
-        let mut dep_secrets = HashSet::new();
-        if let Some(ConnectionInfo::ConnectionParams(params)) = &pb_connection.info {
-            dep_secrets.extend(
-                params
-                    .secret_refs
-                    .values()
-                    .map(|secret_ref| secret_ref.secret_id),
-            );
-        }
-
-        let conn_obj = Self::create_object(
-            &txn,
-            ObjectType::Connection,
-            owner_id,
-            Some(pb_connection.database_id as _),
-            Some(pb_connection.schema_id as _),
-        )
-        .await?;
-        pb_connection.id = conn_obj.oid as _;
-        let connection: connection::ActiveModel = pb_connection.clone().into();
-        Connection::insert(connection).exec(&txn).await?;
-
-        for secret_id in dep_secrets {
-            ObjectDependency::insert(object_dependency::ActiveModel {
-                oid: Set(secret_id as _),
-                used_by: Set(conn_obj.oid),
-                ..Default::default()
-            })
-            .exec(&txn)
-            .await?;
-        }
-
-        txn.commit().await?;
-
-        {
-            // call meta telemetry here to report the connection creation
-            report_event(
-                PbTelemetryEventStage::Unspecified,
-                "connection_create",
-                pb_connection.get_id() as _,
-                {
-                    pb_connection.info.as_ref().and_then(|info| match info {
-                        ConnectionInfo::ConnectionParams(params) => {
-                            Some(params.connection_type().as_str_name().to_owned())
-                        }
-                        _ => None,
-                    })
-                },
-                None,
-                None,
-            );
-        }
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Add,
-                NotificationInfo::Connection(pb_connection),
-            )
-            .await;
-        Ok(version)
-    }
-
     pub async fn drop_connection(
         &self,
         connection_id: ConnectionId,
@@ -1584,560 +1013,6 @@ impl CatalogController {
             )
             .await;
         Ok((version, pb_connection))
-    }
-
-    pub async fn create_view(&self, mut pb_view: PbView) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let owner_id = pb_view.owner as _;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(owner_id, &txn).await?;
-        ensure_object_id(ObjectType::Database, pb_view.database_id as _, &txn).await?;
-        ensure_object_id(ObjectType::Schema, pb_view.schema_id as _, &txn).await?;
-        check_relation_name_duplicate(
-            &pb_view.name,
-            pb_view.database_id as _,
-            pb_view.schema_id as _,
-            &txn,
-        )
-        .await?;
-
-        let view_obj = Self::create_object(
-            &txn,
-            ObjectType::View,
-            owner_id,
-            Some(pb_view.database_id as _),
-            Some(pb_view.schema_id as _),
-        )
-        .await?;
-        pb_view.id = view_obj.oid as _;
-        let view: view::ActiveModel = pb_view.clone().into();
-        View::insert(view).exec(&txn).await?;
-
-        // todo: change `dependent_relations` to `dependent_objects`, which should includes connection and function as well.
-        // todo: shall we need to check existence of them Or let database handle it by FOREIGN KEY constraint.
-        for obj_id in &pb_view.dependent_relations {
-            ObjectDependency::insert(object_dependency::ActiveModel {
-                oid: Set(*obj_id as _),
-                used_by: Set(view_obj.oid),
-                ..Default::default()
-            })
-            .exec(&txn)
-            .await?;
-        }
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend_relation_info(
-                NotificationOperation::Add,
-                PbRelationInfo::View(pb_view),
-            )
-            .await;
-        Ok(version)
-    }
-
-    pub async fn alter_owner(
-        &self,
-        object_type: ObjectType,
-        object_id: ObjectId,
-        new_owner: UserId,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        ensure_user_id(new_owner, &txn).await?;
-
-        let obj = Object::find_by_id(object_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found(object_type.as_str(), object_id))?;
-        if obj.owner_id == new_owner {
-            return Ok(IGNORED_NOTIFICATION_VERSION);
-        }
-        let mut obj = obj.into_active_model();
-        obj.owner_id = Set(new_owner);
-        let obj = obj.update(&txn).await?;
-
-        let mut relations = vec![];
-        match object_type {
-            ObjectType::Database => {
-                let db = Database::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("database", object_id))?;
-
-                txn.commit().await?;
-
-                let version = self
-                    .notify_frontend(
-                        NotificationOperation::Update,
-                        NotificationInfo::Database(ObjectModel(db, obj).into()),
-                    )
-                    .await;
-                return Ok(version);
-            }
-            ObjectType::Schema => {
-                let schema = Schema::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("schema", object_id))?;
-
-                txn.commit().await?;
-
-                let version = self
-                    .notify_frontend(
-                        NotificationOperation::Update,
-                        NotificationInfo::Schema(ObjectModel(schema, obj).into()),
-                    )
-                    .await;
-                return Ok(version);
-            }
-            ObjectType::Table => {
-                let table = Table::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("table", object_id))?;
-
-                // associated source.
-                if let Some(associated_source_id) = table.optional_associated_source_id {
-                    let src_obj = object::ActiveModel {
-                        oid: Set(associated_source_id as _),
-                        owner_id: Set(new_owner),
-                        ..Default::default()
-                    }
-                    .update(&txn)
-                    .await?;
-                    let source = Source::find_by_id(associated_source_id)
-                        .one(&txn)
-                        .await?
-                        .ok_or_else(|| {
-                            MetaError::catalog_id_not_found("source", associated_source_id)
-                        })?;
-                    relations.push(PbRelationInfo::Source(ObjectModel(source, src_obj).into()));
-                }
-
-                // indexes.
-                let (index_ids, mut table_ids): (Vec<IndexId>, Vec<TableId>) =
-                    if table.table_type == TableType::Table {
-                        Index::find()
-                            .select_only()
-                            .columns([index::Column::IndexId, index::Column::IndexTableId])
-                            .filter(index::Column::PrimaryTableId.eq(object_id))
-                            .into_tuple::<(IndexId, TableId)>()
-                            .all(&txn)
-                            .await?
-                            .into_iter()
-                            .unzip()
-                    } else {
-                        (vec![], vec![])
-                    };
-                relations.push(PbRelationInfo::Table(ObjectModel(table, obj).into()));
-
-                // internal tables.
-                let internal_tables: Vec<TableId> = Table::find()
-                    .select_only()
-                    .column(table::Column::TableId)
-                    .filter(
-                        table::Column::BelongsToJobId
-                            .is_in(table_ids.iter().cloned().chain(std::iter::once(object_id))),
-                    )
-                    .into_tuple()
-                    .all(&txn)
-                    .await?;
-                table_ids.extend(internal_tables);
-
-                if !index_ids.is_empty() || !table_ids.is_empty() {
-                    Object::update_many()
-                        .col_expr(
-                            object::Column::OwnerId,
-                            SimpleExpr::Value(Value::Int(Some(new_owner))),
-                        )
-                        .filter(
-                            object::Column::Oid
-                                .is_in(index_ids.iter().cloned().chain(table_ids.iter().cloned())),
-                        )
-                        .exec(&txn)
-                        .await?;
-                }
-
-                if !table_ids.is_empty() {
-                    let table_objs = Table::find()
-                        .find_also_related(Object)
-                        .filter(table::Column::TableId.is_in(table_ids))
-                        .all(&txn)
-                        .await?;
-                    for (table, table_obj) in table_objs {
-                        relations.push(PbRelationInfo::Table(
-                            ObjectModel(table, table_obj.unwrap()).into(),
-                        ));
-                    }
-                }
-                // FIXME: frontend will update index/primary table from cache, requires apply updates of indexes after tables.
-                if !index_ids.is_empty() {
-                    let index_objs = Index::find()
-                        .find_also_related(Object)
-                        .filter(index::Column::IndexId.is_in(index_ids))
-                        .all(&txn)
-                        .await?;
-                    for (index, index_obj) in index_objs {
-                        relations.push(PbRelationInfo::Index(
-                            ObjectModel(index, index_obj.unwrap()).into(),
-                        ));
-                    }
-                }
-            }
-            ObjectType::Source => {
-                let source = Source::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("source", object_id))?;
-                let is_shared = source.is_shared();
-                relations.push(PbRelationInfo::Source(ObjectModel(source, obj).into()));
-
-                // Note: For non-shared source, we don't update their state tables, which
-                // belongs to the MV.
-                if is_shared {
-                    update_internal_tables(
-                        &txn,
-                        object_id,
-                        object::Column::OwnerId,
-                        Value::Int(Some(new_owner)),
-                        &mut relations,
-                    )
-                    .await?;
-                }
-            }
-            ObjectType::Sink => {
-                let sink = Sink::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("sink", object_id))?;
-                relations.push(PbRelationInfo::Sink(ObjectModel(sink, obj).into()));
-
-                update_internal_tables(
-                    &txn,
-                    object_id,
-                    object::Column::OwnerId,
-                    Value::Int(Some(new_owner)),
-                    &mut relations,
-                )
-                .await?;
-            }
-            ObjectType::Subscription => {
-                let subscription = Subscription::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("subscription", object_id))?;
-                relations.push(PbRelationInfo::Subscription(
-                    ObjectModel(subscription, obj).into(),
-                ));
-            }
-            ObjectType::View => {
-                let view = View::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("view", object_id))?;
-                relations.push(PbRelationInfo::View(ObjectModel(view, obj).into()));
-            }
-            _ => unreachable!("not supported object type: {:?}", object_type),
-        };
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::RelationGroup(PbRelationGroup {
-                    relations: relations
-                        .into_iter()
-                        .map(|relation| PbRelation {
-                            relation_info: Some(relation),
-                        })
-                        .collect(),
-                }),
-            )
-            .await;
-        Ok(version)
-    }
-
-    pub async fn alter_schema(
-        &self,
-        object_type: ObjectType,
-        object_id: ObjectId,
-        new_schema: SchemaId,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        ensure_object_id(ObjectType::Schema, new_schema, &txn).await?;
-
-        let obj = Object::find_by_id(object_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found(object_type.as_str(), object_id))?;
-        if obj.schema_id == Some(new_schema) {
-            return Ok(IGNORED_NOTIFICATION_VERSION);
-        }
-        let database_id = obj.database_id.unwrap();
-
-        let mut relations = vec![];
-        match object_type {
-            ObjectType::Table => {
-                let table = Table::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("table", object_id))?;
-                check_relation_name_duplicate(&table.name, database_id, new_schema, &txn).await?;
-                let (associated_src_id, table_type) =
-                    (table.optional_associated_source_id, table.table_type);
-
-                let mut obj = obj.into_active_model();
-                obj.schema_id = Set(Some(new_schema));
-                let obj = obj.update(&txn).await?;
-                relations.push(PbRelationInfo::Table(ObjectModel(table, obj).into()));
-
-                // associated source.
-                if let Some(associated_source_id) = associated_src_id {
-                    let src_obj = object::ActiveModel {
-                        oid: Set(associated_source_id as _),
-                        schema_id: Set(Some(new_schema)),
-                        ..Default::default()
-                    }
-                    .update(&txn)
-                    .await?;
-                    let source = Source::find_by_id(associated_source_id)
-                        .one(&txn)
-                        .await?
-                        .ok_or_else(|| {
-                            MetaError::catalog_id_not_found("source", associated_source_id)
-                        })?;
-                    relations.push(PbRelationInfo::Source(ObjectModel(source, src_obj).into()));
-                }
-
-                // indexes.
-                let (index_ids, (index_names, mut table_ids)): (
-                    Vec<IndexId>,
-                    (Vec<String>, Vec<TableId>),
-                ) = if table_type == TableType::Table {
-                    Index::find()
-                        .select_only()
-                        .columns([
-                            index::Column::IndexId,
-                            index::Column::Name,
-                            index::Column::IndexTableId,
-                        ])
-                        .filter(index::Column::PrimaryTableId.eq(object_id))
-                        .into_tuple::<(IndexId, String, TableId)>()
-                        .all(&txn)
-                        .await?
-                        .into_iter()
-                        .map(|(id, name, t_id)| (id, (name, t_id)))
-                        .unzip()
-                } else {
-                    (vec![], (vec![], vec![]))
-                };
-
-                // internal tables.
-                let internal_tables: Vec<TableId> = Table::find()
-                    .select_only()
-                    .column(table::Column::TableId)
-                    .filter(
-                        table::Column::BelongsToJobId
-                            .is_in(table_ids.iter().cloned().chain(std::iter::once(object_id))),
-                    )
-                    .into_tuple()
-                    .all(&txn)
-                    .await?;
-                table_ids.extend(internal_tables);
-
-                if !index_ids.is_empty() || !table_ids.is_empty() {
-                    for index_name in index_names {
-                        check_relation_name_duplicate(&index_name, database_id, new_schema, &txn)
-                            .await?;
-                    }
-
-                    Object::update_many()
-                        .col_expr(
-                            object::Column::SchemaId,
-                            SimpleExpr::Value(Value::Int(Some(new_schema))),
-                        )
-                        .filter(
-                            object::Column::Oid
-                                .is_in(index_ids.iter().cloned().chain(table_ids.iter().cloned())),
-                        )
-                        .exec(&txn)
-                        .await?;
-                }
-
-                if !table_ids.is_empty() {
-                    let table_objs = Table::find()
-                        .find_also_related(Object)
-                        .filter(table::Column::TableId.is_in(table_ids))
-                        .all(&txn)
-                        .await?;
-                    for (table, table_obj) in table_objs {
-                        relations.push(PbRelationInfo::Table(
-                            ObjectModel(table, table_obj.unwrap()).into(),
-                        ));
-                    }
-                }
-                if !index_ids.is_empty() {
-                    let index_objs = Index::find()
-                        .find_also_related(Object)
-                        .filter(index::Column::IndexId.is_in(index_ids))
-                        .all(&txn)
-                        .await?;
-                    for (index, index_obj) in index_objs {
-                        relations.push(PbRelationInfo::Index(
-                            ObjectModel(index, index_obj.unwrap()).into(),
-                        ));
-                    }
-                }
-            }
-            ObjectType::Source => {
-                let source = Source::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("source", object_id))?;
-                check_relation_name_duplicate(&source.name, database_id, new_schema, &txn).await?;
-                let is_shared = source.is_shared();
-
-                let mut obj = obj.into_active_model();
-                obj.schema_id = Set(Some(new_schema));
-                let obj = obj.update(&txn).await?;
-                relations.push(PbRelationInfo::Source(ObjectModel(source, obj).into()));
-
-                // Note: For non-shared source, we don't update their state tables, which
-                // belongs to the MV.
-                if is_shared {
-                    update_internal_tables(
-                        &txn,
-                        object_id,
-                        object::Column::SchemaId,
-                        Value::Int(Some(new_schema)),
-                        &mut relations,
-                    )
-                    .await?;
-                }
-            }
-            ObjectType::Sink => {
-                let sink = Sink::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("sink", object_id))?;
-                check_relation_name_duplicate(&sink.name, database_id, new_schema, &txn).await?;
-
-                let mut obj = obj.into_active_model();
-                obj.schema_id = Set(Some(new_schema));
-                let obj = obj.update(&txn).await?;
-                relations.push(PbRelationInfo::Sink(ObjectModel(sink, obj).into()));
-
-                update_internal_tables(
-                    &txn,
-                    object_id,
-                    object::Column::SchemaId,
-                    Value::Int(Some(new_schema)),
-                    &mut relations,
-                )
-                .await?;
-            }
-            ObjectType::Subscription => {
-                let subscription = Subscription::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("subscription", object_id))?;
-                check_relation_name_duplicate(&subscription.name, database_id, new_schema, &txn)
-                    .await?;
-
-                let mut obj = obj.into_active_model();
-                obj.schema_id = Set(Some(new_schema));
-                let obj = obj.update(&txn).await?;
-                relations.push(PbRelationInfo::Subscription(
-                    ObjectModel(subscription, obj).into(),
-                ));
-            }
-            ObjectType::View => {
-                let view = View::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("view", object_id))?;
-                check_relation_name_duplicate(&view.name, database_id, new_schema, &txn).await?;
-
-                let mut obj = obj.into_active_model();
-                obj.schema_id = Set(Some(new_schema));
-                let obj = obj.update(&txn).await?;
-                relations.push(PbRelationInfo::View(ObjectModel(view, obj).into()));
-            }
-            ObjectType::Function => {
-                let function = Function::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("function", object_id))?;
-
-                let mut pb_function: PbFunction = ObjectModel(function, obj).into();
-                pb_function.schema_id = new_schema as _;
-                check_function_signature_duplicate(&pb_function, &txn).await?;
-
-                object::ActiveModel {
-                    oid: Set(object_id),
-                    schema_id: Set(Some(new_schema)),
-                    ..Default::default()
-                }
-                .update(&txn)
-                .await?;
-
-                txn.commit().await?;
-                let version = self
-                    .notify_frontend(
-                        NotificationOperation::Update,
-                        NotificationInfo::Function(pb_function),
-                    )
-                    .await;
-                return Ok(version);
-            }
-            ObjectType::Connection => {
-                let connection = Connection::find_by_id(object_id)
-                    .one(&txn)
-                    .await?
-                    .ok_or_else(|| MetaError::catalog_id_not_found("connection", object_id))?;
-
-                let mut pb_connection: PbConnection = ObjectModel(connection, obj).into();
-                pb_connection.schema_id = new_schema as _;
-                check_connection_name_duplicate(&pb_connection, &txn).await?;
-
-                object::ActiveModel {
-                    oid: Set(object_id),
-                    schema_id: Set(Some(new_schema)),
-                    ..Default::default()
-                }
-                .update(&txn)
-                .await?;
-
-                txn.commit().await?;
-                let version = self
-                    .notify_frontend(
-                        NotificationOperation::Update,
-                        NotificationInfo::Connection(pb_connection),
-                    )
-                    .await;
-                return Ok(version);
-            }
-            _ => unreachable!("not supported object type: {:?}", object_type),
-        }
-
-        txn.commit().await?;
-        let version = self
-            .notify_frontend(
-                Operation::Update,
-                Info::RelationGroup(PbRelationGroup {
-                    relations: relations
-                        .into_iter()
-                        .map(|relation_info| PbRelation {
-                            relation_info: Some(relation_info),
-                        })
-                        .collect_vec(),
-                }),
-            )
-            .await;
-        Ok(version)
     }
 
     pub async fn comment_on(&self, comment: PbComment) -> MetaResult<NotificationVersion> {
@@ -2453,242 +1328,6 @@ impl CatalogController {
             },
             version,
         ))
-    }
-
-    async fn alter_database_name(
-        &self,
-        database_id: DatabaseId,
-        name: &str,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        check_database_name_duplicate(name, &txn).await?;
-
-        let active_model = database::ActiveModel {
-            database_id: Set(database_id),
-            name: Set(name.to_owned()),
-        };
-        let database = active_model.update(&txn).await?;
-
-        let obj = Object::find_by_id(database_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("database", database_id))?;
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::Database(ObjectModel(database, obj).into()),
-            )
-            .await;
-        Ok(version)
-    }
-
-    async fn alter_schema_name(
-        &self,
-        schema_id: SchemaId,
-        name: &str,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-
-        let obj = Object::find_by_id(schema_id)
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("schema", schema_id))?;
-        check_schema_name_duplicate(name, obj.database_id.unwrap(), &txn).await?;
-
-        let active_model = schema::ActiveModel {
-            schema_id: Set(schema_id),
-            name: Set(name.to_owned()),
-        };
-        let schema = active_model.update(&txn).await?;
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::Schema(ObjectModel(schema, obj).into()),
-            )
-            .await;
-        Ok(version)
-    }
-
-    pub async fn alter_name(
-        &self,
-        object_type: ObjectType,
-        object_id: ObjectId,
-        object_name: &str,
-    ) -> MetaResult<NotificationVersion> {
-        if object_type == ObjectType::Database {
-            return self.alter_database_name(object_id as _, object_name).await;
-        } else if object_type == ObjectType::Schema {
-            return self.alter_schema_name(object_id as _, object_name).await;
-        }
-
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let obj: PartialObject = Object::find_by_id(object_id)
-            .into_partial_model()
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found(object_type.as_str(), object_id))?;
-        assert_eq!(obj.obj_type, object_type);
-        check_relation_name_duplicate(
-            object_name,
-            obj.database_id.unwrap(),
-            obj.schema_id.unwrap(),
-            &txn,
-        )
-        .await?;
-
-        // rename relation.
-        let (mut to_update_relations, old_name) =
-            rename_relation(&txn, object_type, object_id, object_name).await?;
-        // rename referring relation name.
-        to_update_relations.extend(
-            rename_relation_refer(&txn, object_type, object_id, object_name, &old_name).await?,
-        );
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::RelationGroup(PbRelationGroup {
-                    relations: to_update_relations,
-                }),
-            )
-            .await;
-
-        Ok(version)
-    }
-
-    pub async fn alter_swap_rename(
-        &self,
-        object_type: ObjectType,
-        object_id: ObjectId,
-        dst_object_id: ObjectId,
-    ) -> MetaResult<NotificationVersion> {
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-        let dst_name: String = match object_type {
-            ObjectType::Table => Table::find_by_id(dst_object_id)
-                .select_only()
-                .column(table::Column::Name)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| {
-                    MetaError::catalog_id_not_found(object_type.as_str(), dst_object_id)
-                })?,
-            ObjectType::Source => Source::find_by_id(dst_object_id)
-                .select_only()
-                .column(source::Column::Name)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| {
-                    MetaError::catalog_id_not_found(object_type.as_str(), dst_object_id)
-                })?,
-            ObjectType::Sink => Sink::find_by_id(dst_object_id)
-                .select_only()
-                .column(sink::Column::Name)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| {
-                    MetaError::catalog_id_not_found(object_type.as_str(), dst_object_id)
-                })?,
-            ObjectType::View => View::find_by_id(dst_object_id)
-                .select_only()
-                .column(view::Column::Name)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| {
-                    MetaError::catalog_id_not_found(object_type.as_str(), dst_object_id)
-                })?,
-            ObjectType::Subscription => Subscription::find_by_id(dst_object_id)
-                .select_only()
-                .column(subscription::Column::Name)
-                .into_tuple()
-                .one(&txn)
-                .await?
-                .ok_or_else(|| {
-                    MetaError::catalog_id_not_found(object_type.as_str(), dst_object_id)
-                })?,
-            _ => {
-                return Err(MetaError::permission_denied(format!(
-                    "swap rename not supported for object type: {:?}",
-                    object_type
-                )));
-            }
-        };
-
-        // rename relations.
-        let (mut to_update_relations, src_name) =
-            rename_relation(&txn, object_type, object_id, &dst_name).await?;
-        let (to_update_relations2, _) =
-            rename_relation(&txn, object_type, dst_object_id, &src_name).await?;
-        to_update_relations.extend(to_update_relations2);
-        // rename referring relation name.
-        to_update_relations.extend(
-            rename_relation_refer(&txn, object_type, object_id, &dst_name, &src_name).await?,
-        );
-        to_update_relations.extend(
-            rename_relation_refer(&txn, object_type, dst_object_id, &src_name, &dst_name).await?,
-        );
-
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend(
-                NotificationOperation::Update,
-                NotificationInfo::RelationGroup(PbRelationGroup {
-                    relations: to_update_relations,
-                }),
-            )
-            .await;
-
-        Ok(version)
-    }
-
-    pub async fn alter_non_shared_source(
-        &self,
-        pb_source: PbSource,
-    ) -> MetaResult<NotificationVersion> {
-        let source_id = pb_source.id as SourceId;
-        let inner = self.inner.write().await;
-        let txn = inner.db.begin().await?;
-
-        let original_version: i64 = Source::find_by_id(source_id)
-            .select_only()
-            .column(source::Column::Version)
-            .into_tuple()
-            .one(&txn)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("source", source_id))?;
-        if original_version + 1 != pb_source.version as i64 {
-            return Err(MetaError::permission_denied(
-                "source version is stale".to_owned(),
-            ));
-        }
-
-        let source: source::ActiveModel = pb_source.clone().into();
-        source.update(&txn).await?;
-        txn.commit().await?;
-
-        let version = self
-            .notify_frontend_relation_info(
-                NotificationOperation::Update,
-                PbRelationInfo::Source(pb_source),
-            )
-            .await;
-        Ok(version)
     }
 
     pub async fn find_creating_streaming_job_ids(
@@ -3118,34 +1757,4 @@ impl CatalogControllerInner {
             .await?;
         Ok(table_ids)
     }
-}
-
-async fn update_internal_tables(
-    txn: &DatabaseTransaction,
-    object_id: i32,
-    column: object::Column,
-    new_value: Value,
-    relations_to_notify: &mut Vec<PbRelationInfo>,
-) -> MetaResult<()> {
-    let internal_tables = get_internal_tables_by_id(object_id, txn).await?;
-
-    if !internal_tables.is_empty() {
-        Object::update_many()
-            .col_expr(column, SimpleExpr::Value(new_value))
-            .filter(object::Column::Oid.is_in(internal_tables.clone()))
-            .exec(txn)
-            .await?;
-
-        let table_objs = Table::find()
-            .find_also_related(Object)
-            .filter(table::Column::TableId.is_in(internal_tables))
-            .all(txn)
-            .await?;
-        for (table, table_obj) in table_objs {
-            relations_to_notify.push(PbRelationInfo::Table(
-                ObjectModel(table, table_obj.unwrap()).into(),
-            ));
-        }
-    }
-    Ok(())
 }
